+++ src/sdk/__mocks__/ControlPanelController.ts
@@ -1,0 +1,4 @@
+
+export const ControlPanelController = {
+    controller: {}
+}
