+++ src/sdk/__mocks__/MapController.ts
@@ -1,0 +1,6 @@
+
+export const MapController = {
+    controller: {
+        updateOrbsMask: () => true,
+    }
+}
