+++ src/sdk/__mocks__/XpDropController.ts
@@ -1,0 +1,7 @@
+
+export const XpDropController = {
+    controller: {
+        tick: () => true,
+        registerXpDrop: () => true,
+    }
+}
