+++ test/simulations/InfernoRanger.test.ts
@@ -1,3 +1,7 @@
+jest.mock("../../src/sdk/ControlPanelController");
+jest.mock("../../src/sdk/MapController");
+jest.mock("../../src/sdk/XpDropController");
+
 import { JalXil } from '../../src/content/inferno/js/mobs/JalXil';
 import { Player } from '../../src/sdk/Player';
 import { World } from '../../src/sdk/World';
@@ -6,10 +10,22 @@ import { TwistedBow } from "../../src/content/weapons/TwistedBow";
 import { Region } from "../../src/sdk/Region";
 import { Random } from "../../src/sdk/Random";
 import { Viewport } from "../../src/sdk/Viewport";
+import { Settings } from '../../src/sdk/Settings';
 
-import setup from "./common";
+jest
+  .spyOn(document, "getElementById")
+  .mockImplementation((elementId: string) => {
+    const c = document.createElement("canvas");
+    c.ariaLabel = elementId;
+    return c;
+  });
 
-setup();
+Random.setRandom(() => {
+  Random.memory = (Random.memory + 13.37) % 180;
+  return Math.abs(Math.sin(Random.memory * 0.0174533));
+});
+
+Settings.readFromStorage();
 
 export class TestRegion60x60 extends Region {
   get width(): number {
